+++ CHANGES.md
@@ -18,8 +18,8 @@ You can upgrade from 2.2.0 without worries.
 * When validating with dry-validation, we now pass a symbolized hash. We also replaced `Dry::Validation::Form` with `Schema` which won't coerce values where it shouldn't.
 * [private] `Group#call` API now is: `call(form, errors)`.
 * Removed `Form#valid?`.
-
 * In `:if` for validation groups, you now get a hash of result objects, not just true/false.
+* Allow adding a custom error AFTER validate has been already called
 
 
 ## 2.2.4